yolo_api/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
ultralytics
